feat: implement TOML config parsing
Fetches app configurations from config.toml and keys.toml files.
Currently only works if the executable is present in the repo folder or
one of the subfolders, as the algorithm recursively looks for the
project root. Support for environmental variables has not yet
been implemented.

Application-specific settings don't ovewrite defaults yet, even if
uncommented.

Type mismatches are not handled as of this commit, so the program will
crash if conversions fail.